refactor: share groupReposByLanguage
